% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milli_quarto.R
\name{milli_quarto}
\alias{milli_quarto}
\title{milli_quarto}
\usage{
milli_quarto()
}
\value{
Quarto header
}
\description{
milli_quarto provides a clean set-up for all PDF documents and can be copy-pasted
}
