% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millis_mode.R
\name{millis_mode}
\alias{millis_mode}
\title{millis_mode}
\usage{
millis_mode(x)
}
\value{
Mode function
}
\description{
Mode function: all the credits go to stack overflow!
 https://stackoverflow.com/questions/2547402/how-to-find-the-statistical-mode
}
