% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millis_theme.R
\name{millis_theme}
\alias{millis_theme}
\title{millis_theme}
\usage{
millis_theme(
  base_size = 11,
  base_family = "Century Gothic",
  base_line_size = base_size/120,
  base_rect_size = base_size/120
)
}
\value{
Theme to match the fonts
}
\description{
This theme matches the quarto header
 I USE GGPLOT AND BUILD ON THEME MINIMAL
}
