% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millis_unload.R
\name{millis_unload}
\alias{millis_unload}
\title{millis_unload}
\usage{
millis_unload()
}
\value{
Unloading Packages
}
\description{
This function unloads most packages.
 Basics of R are not unloaded to avoid a crash.
}
